Final Commit
Wrote a basic final report discussing some reasoning behind deviations
from original plan (mainly the absence of the ability of the AI to
"forget" which concealed piece is which).  Also included references and
screenshots
displaying the operation of the AI.
Updated the status summary (all remaining criteria marked Done (green),
except for the "ability to forget" criterion, which was marked abandoned
(red) with a reference to the final report for an explanation.
Added a README describing the basic rules of
Schrodinger's Chess with instructions on starting and
playing a game.